refactor(download): Wave 4 — FileDownloader, ModelCache, ProgressReporter; DownloadUtils becomes a façade (#765)
Behavior-preserving extraction; the Wave 1 oracle (filter fixtures,
progress sequences, offline suite) plus the resume/artifact/cancellation
suites ride along untouched. DownloadUtils: 1220 → ~650 lines.

- FileDownloader (Shared/Download/): the per-file unit — ensure(file:
  from:at:) = skip-if-present / create-empty / authorized download with
  RetryPolicy + byte-range resume (#757 transport, moved verbatim:
  streamDownload, StreamingDownloadDelegate, resume helpers,
  validateDownloadedArtifact) / atomic move. Both downloadRepo and
  downloadSubdirectory loops now call it.
- ModelCache (Shared/Download/): allModelsExist/missingModels (offline
  path), verifyModelsPresent (post-download pass),
  validateCompiledModelLayout (pre-compile checks, moved verbatim),
  purgeCorruptedCache (loadModels recovery; the cancellation-is-not-
  corruption guard stays at the loadModels call site), and
  createDirectoryRobustly.
- ProgressReporter (Shared/Download/): weighted phases with the
  documented invariant 'monotonic fraction of THIS operation; phase says
  what is happening'. Repo loads declare download 0.5 / compile 0.5
  (byte-weighted, exact old emissions); subdirectory declares 1.0 (the
  #769 math, absorbed as ProgressReporter.downloadFraction).
- Test-pinned forwards kept on DownloadUtils: downloadFileWithRetry,
  resumeValidatorURL, validateDownloadedArtifact,
  subdirectoryProgressFraction, isRetryableDownloadError,
  isCancellationError.
- Deliberate micro-deltas (ledger): downloadSubdirectory's parent-dir
  creation now uses createDirectoryRobustly (was plain createDirectory —
  strictly more capable); per-file transport logs (Resuming/retry
  warnings) move to the FileDownloader log category.

Verified: cache layout byte-identical — fresh silero download on a clean
main build vs this branch, same paths + sizes; both loads succeed.
